% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{match}
\usage{
match(.data, pattern)
}
\arguments{
\item{.data}{a data frame or tibble.}

\item{pattern}{a character string that is used to find the columns of interest. It can be a regular expression.}
}
\description{
`match` extracts columns in a data frame that match a specified regex string.
}
