% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{match}
\usage{
match(.data, pattern, ignore.case = FALSE)
}
\arguments{
\item{.data}{a data frame or tibble.}

\item{pattern}{a character string that is used to find the columns of interest. It can be a regular expression.}

\item{ignore.case}{a logical scalar that specifies whether the case of the string should be ignored when matching. Defaults to `FALSE`.}
}
\description{
`match` extracts columns in a data frame that match a specified regex string.
}
